move call_RCCL routine StandaloneUtils, create .cpp file for StandaloneUtils so that it can be included in several tests
