+++ test/AllReduceTests.cpp
@@ -4,7 +4,7 @@
  * See LICENSE.txt for license information
  ************************************************************************/
 #include "TestBed.hpp"
-
+#include "StandaloneUtils.hpp"
 namespace RcclUnitTesting
 {
   TEST(AllReduce, OutOfPlace)
